week7/abilities_django/requirements.txt
Django==1.10
djangorestframework==3.4.5
psycopg2==2.6.2
